TradingBot/tradingbot2/requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
python-dotenv>=0.19.0
pydantic>=1.10.0
requests>=2.28.0


